Fixed Create Product error in CoreAdmin

+++ Src/ElectronicsShop-service/Dtos/ProductDto.cs
@@ -17,8 +17,8 @@ public class ProductDto : BaseDto
     public string? description { get; set; }
     public byte[]? image { get; set; }
 
-    public float Rating { get; set; }
-    public float price { get; set; }
+    public decimal Rating { get; set; }
+    public decimal price { get; set; }
     public int? quantity { get; set; }
     public string? status { get; set; }
